fixes, add lesskey and finally make mg C-\ as undo
Previous mg attempts to set this failed with segmentation violation as
the required sequence "\^\" is difficult to parse.  Examination of the
code find it groups as '\^'+'\"' which escapes the closing quote
(comments in the code about this point). I noticed that "\nnn" octal
sequence is handled so the fix is to replace "\^\" with "\034" which
works; describe-bindings now shows:   C-\ undo

lesskey adds d or D to search for diff and brackets [] as next/previous file
from: http://chneukirchen.org/blog/archive/2015/02/six-hacks-for-less-1.html
this url was mentioned in the post: https://www.dragonflydigest.com/2018/06/10/21353.html

Signed-off-by: Christopher Hall <hsw@ms2.hinet.net>